# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "To-Do List API",
    "description": "Documentação Feita para aula, abaixo terão todas as rotas utilizadas nos dias de aula.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/tasks": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Detalhes da tarefa",
            "schema": {
              "$ref": "#/definitions/addTask"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/tasks/active": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/tasks/completed": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/tasks/todos/{id}": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/register": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Registro de usuário",
            "schema": {
              "$ref": "#/definitions/register"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Login",
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "addTask": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "Titulo da tarefa"
        },
        "completed": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "login": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Nome do usuário"
        },
        "password": {
          "type": "string",
          "example": "Senha do usuário"
        }
      }
    },
    "register": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Nome do usuário"
        },
        "password": {
          "type": "string",
          "example": "Senha do usuário"
        },
        "image": {
          "type": "string",
          "example": "foto do usuário (opcional)"
        }
      }
    }
  }
}